Do not attempt to auto-discover ingress in tests if 'ingressendpoint' flag is set
Previously if auto-discovery of the public ingress failed then the
tests would fail, even if the ingress address was being overriden
using the '-ingressendpoint' flag. This change makes it so that
ingress discovery is skipped if the '-ingressendpoint' flag is set.

I needed this change in order to run the tests using a node port for
ingress in a local kubernetes instance. It is possible my cluster
isn't set up exactly as expected however I think it makes sense for
discovery to be skipped when the ingress endpoint has been set
manually.